move initTest to setUp in TestBase add FriendshipTest move friendshipInviteTest and revokeFriendshipTest to FriendshipTest fix deletePost in PostCard fix addToFriends in PersonPage
