+++ src/main/java/pages/NewTopicCard.java
@@ -1,11 +1,9 @@
 package pages;
 
-import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
 import org.openqa.selenium.By;
 
-import static com.codeborne.selenide.Condition.appear;
-import static com.codeborne.selenide.Condition.appears;
+import static com.codeborne.selenide.Condition.*;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.$x;
 import static java.lang.String.format;
@@ -20,13 +18,15 @@ public class NewTopicCard {
     private static final By RESTORE_DRAFT_WINDOW = xpath(".//*[contains(@id, \"QuitConfirmationForm\")]");
     private static final By SHARE_BUTTON = xpath(".//*[@data-action=\"submit\"]");
     private static final By CLOSE_VIDEO = xpath(".//*[@class='ic media-layer_close_ico']");
+    private static final By CANCEL_RESTORE = xpath(".//*[@data-js-cancel='1']");
     public static final By CREATE_TOPIC_CARD = cssSelector("div#hook_Block_MediaTopicLayerBody");
     private SelenideElement element;
 
     public NewTopicCard() {
         element = $(CREATE_TOPIC_CARD).scrollTo();
         if ($(RESTORE_DRAFT_WINDOW).isDisplayed()) {
-            $x(".//*[@data-js-cancel='1']").click();
+            $(CANCEL_RESTORE).waitUntil(visible, 5000, 400);
+            $(CANCEL_RESTORE).click();
         }
     }
 
@@ -37,13 +37,15 @@ public class NewTopicCard {
 
     public NewTopicCard addVideoWithTitle(String title) {
         element.$(ADD_VIDEO_BUTTON).click();
-        $x(format(VIDEO_WITH_TITLE, title)).waitUntil(Condition.visible, 5000, 400);
+        $x(format(VIDEO_WITH_TITLE, title)).waitUntil(appear, 5000, 400);
+        $x(format(VIDEO_WITH_TITLE, title)).hover();
         $x(format(VIDEO_WITH_TITLE, title)).click();
+        $x(format(VIDEO_WITH_TITLE, title)).waitUntil(disappear, 5000, 400);
         return this;
     }
 
     public GroupPage sharePost() {
-        element.$(SHARE_BUTTON).waitUntil(appear, 5000, 400);
+        element.$(SHARE_BUTTON).waitUntil(visible, 5000, 400);
         element.$(SHARE_BUTTON).click();
         return new GroupPage();
     }
