+++ src/test/java/tests/TestBase.java
@@ -1,9 +1,13 @@
 package tests;
 
 import com.codeborne.selenide.Configuration;
+import model.BotFactory;
 import model.TestBot;
 import org.junit.jupiter.api.BeforeAll;
+import pages.LoginPage;
+import pages.PersonPage;
 
+import static com.codeborne.selenide.Configuration.baseUrl;
 import static com.codeborne.selenide.Configuration.browser;
 import static org.apache.commons.lang3.RandomStringUtils.randomAlphanumeric;
 
@@ -12,12 +16,12 @@ public abstract class TestBase {
 
     @BeforeAll
     public static void setUp() {
-        init();
-    }
-
-    private static void init() {
         browser = "chrome";
         Configuration.timeout = 5000;
+        baseUrl = "https://ok.ru/";
+        bot = BotFactory.getOkBot();
+        new PersonPage().openHomePage();
+        new LoginPage().logIn(bot);
     }
 
     protected static String getRandomString() {
